Combine shape/mask; reference frame to calibration.
