+++ nion/data/DataAndMetadata.py
@@ -13,7 +13,6 @@ import warnings
 import numpy
 from nion.data import Calibration
 from nion.data import Image
-from nion.data import ReferenceFrame
 
 _ = gettext.gettext
 
@@ -367,10 +366,10 @@ class DataMetadata:
         return False
 
     @property
-    def reference_frame_2d(self) -> ReferenceFrame.ReferenceFrame2D:
+    def reference_frame_2d(self) -> Calibration.ReferenceFrame2D:
         assert len(self.data_shape) == 2 and self.datum_dimension_count == 2
-        return ReferenceFrame.ReferenceFrame2D(ReferenceFrame.ReferenceFrameAxis(self.dimensional_calibrations[0], self.data_shape[0]),
-                                               ReferenceFrame.ReferenceFrameAxis(self.dimensional_calibrations[1], self.data_shape[1]))
+        return Calibration.ReferenceFrame2D(Calibration.ReferenceFrameAxis(self.dimensional_calibrations[0], self.data_shape[0]),
+                                            Calibration.ReferenceFrameAxis(self.dimensional_calibrations[1], self.data_shape[1]))
 
     def __get_size_str(self, data_shape: typing.Sequence[int], is_spatial: bool = False) -> str:
         spatial_shape_str = " x ".join([str(d) for d in data_shape])
@@ -802,7 +801,7 @@ class DataAndMetadata:
         return self.__data_metadata.is_datum_2d
 
     @property
-    def reference_frame_2d(self) -> ReferenceFrame.ReferenceFrame2D:
+    def reference_frame_2d(self) -> Calibration.ReferenceFrame2D:
         return self.__data_metadata.reference_frame_2d
 
     @property
